\name{science_codes}
\alias{science_codes}
\title{Science and Technology Dataset Codes}
\usage{
science_codes()
}
\value{
Named character vector of dataset codes
}
\description{
Returns a named vector of Eurostat regional science and technology
dataset codes covering R&D expenditure, personnel, patents, HRST,
and high-tech employment at NUTS 2 level.
}
\examples{
science_codes()
names(science_codes())
}
