\name{ssa_codes}
\alias{ssa_codes}
\title{Sub-Saharan Africa DHS Country Codes}
\description{
Returns a character vector of 2-letter DHS country codes for Sub-Saharan
Africa. These are the DHS-assigned codes (which match ISO 3166-1 alpha-2
for most countries).
}
\usage{
ssa_codes()
}
\value{
Character vector of 44 DHS country codes for Sub-Saharan Africa.
}
\examples{
\dontrun{
ssa_codes()
length(ssa_codes())
}
}
\keyword{datasets}
